refactor: Remove platform_ref inheritance from saga definitions
Delete the platform_ref inheritance system including override API,
platform sync, similarity matching, version pinning, and fallback
resolution. Simplify Definition struct by removing PlatformRef,
OverrideReason, PlatformVersionAtOverride, ResolvedScript, and
UsedPlatformFallback fields. Script is now required directly on
all saga definitions.

- Delete 12 files (override_api, similarity, migrate_platform_ref,
  fallback_resolution, platform_sync, version_pinning + tests)
- Remove platform_ref columns from SQL queries and scan functions
- Simplify CreateDraft INSERT and activation validation
- Move ComputeScriptHash to registry.go, humanizeName/regex to seeder.go
- Update orchestrators to use Script instead of ResolvedScript
- Remove PlatformSagaVersionID from shared SagaInstance model
- Add migration to drop platform_ref columns and set script NOT NULL
- Update test DDL to match simplified schema
