+++ services/reference-data/saga/grpc_handler_validation.go
@@ -104,12 +104,7 @@ func (h *RegistryHandler) ValidateSagaDraft(
 		}, nil
 	}
 
-	// Use ResolvedScript for platform-ref sagas (Script is empty, script comes from platform)
-	validationScript := saga.Script
-	if validationScript == "" {
-		validationScript = saga.ResolvedScript
-	}
-	result, err := h.validator.ValidateDraft(ctx, id, validationScript)
+	result, err := h.validator.ValidateDraft(ctx, id, saga.Script)
 	if err != nil {
 		return nil, status.Errorf(codes.Internal, "validation error: %v", err)
 	}
